pyproject.toml
[project]
name = "influx-db-demo"
version = "0.1.0"
description = "Security-aware FastAPI + InfluxDB (v2) demo with Docker and tests"
requires-python = ">=3.11"
dependencies = [
  "bcrypt<5",
  "fastapi>=0.110",
  "httpx>=0.27",
  "influxdb-client>=1.48.0",
  "itsdangerous>=2.2.0",
  "jinja2>=3.1",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.2.1",
  "pyjwt>=2.8.0",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
